Update update System

+++ src/getRepo.sh
@@ -1,7 +1,12 @@
 #!/usr/bin/bash
 
 if [ -d "RaspiOpener" ]; then
-  rm -r RaspiOpener
+  cd RaspiOpener
+  originUrl=$(git remote get-url origin)
+  cd ..
+  if [ $1 != originUrl ]; then
+    rm -r RaspiOpener
+  fi
 fi
 
 git clone $1
